fix : /components/chat/clientSocket move event handler construction from clientSocket.ts@setIncommingMessageHandler to utils.ts@constructIncommingMessageHandler
